+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-## [6.1.3] 2024-05-03
+## [6.5.1] 2024-05-20
+
+### Changed
+
+- [TD-6584] Reduce the number of database queries when requesting metadata
+            for multiple data structure versions through GraphQL
+
+## [6.5.0] 2024-04-30
 
 - [TD-6640] Update td-core-lib to show all error in elastic search